Changed to new type used in libtaquinsolve

+++ Animation/Cat/Cat.cc
@@ -53,7 +53,7 @@ void Cat::reset_puzzle()
     std::string texture_name = "/Animate/data/Cat/" + std::to_string(this->texture_index++) + ".jpg";
     this->texture_index %= 6;
 
-    std::vector<size_t> initial_position = taquin_generate_vector(this->grid_size);
+    std::vector<uint8_t> initial_position = taquin_generate_vector(this->grid_size);
     this->move_sequence = taquin_solve(initial_position, this->grid_size);
 
     //Initialise cat tiles
